+++ R/manhattan.R
@@ -9,7 +9,7 @@
 #'
 #'
 #' @param df Dataframe or a list of dataframes (required columns are \code{CHROM,POS,P}), in upper- or lowercase) of association results.
-#' @param ntop An integer, number of datasets (GWASes) to show on the top plot
+#' @param ntop An integer, number of datasets (GWAS results) to show on the top plot
 #' @param chr A string or integer, the chromosome to plot (i.e. chr15), only required if the input dataframe contains results from more than one chromosome
 #' @param title A string to set the plot title
 #' @param color A string or a vector of strings, for setting the color of the datapoints on the plot
@@ -24,10 +24,10 @@
 #' @param sign_thresh_color A string or vector of strings to set the color/s of the significance threshold/s
 #' @param highlight_genes A string or vector of strings, gene or genes to highlight at the bottom of the plot
 #' @param highlight_genes_ypos An integer, controlling where on the y-axis the highlighted genes are placed (default value is 1)
-#' @param highlight_genes_color A string, color for the highlighted genes (default: green)
+#' @param highlight_genes_color A string, color for the highlighted genes (default: darkred)
 #' @param xmin,xmax Integer, setting the chromosomal range to display on the x-axis
 #' @param ymin,ymax Integer, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})
-#' @param legend_labels A string or vector of strings representing legend labels for the input dataset's
+#' @param legend_labels A string or vector of strings representing legend labels for the input datasets
 #' @param legend_name A string, use to change the name of the legend (default: None)
 #' @param legend_position A string, top,bottom,left or right
 #' @param title_text_size A number, size of the plot title (default: 13)
@@ -37,7 +37,7 @@
 #' @param legend_text_size A number, size of the legend text
 #' @param legend_nrow An integer, sets the number of rows allowed for the legend labels
 #' @param protein_coding_only A logical scalar, if TRUE, only protein coding genes are used for annotation
-#' @param region_size An integer (default = 1000000) (or a string represented as 100kb or 1MB) indicating the window size for variant labeling. Increase this number for sparser annotation and decrease for denser annotation.
+#' @param region_size An integer (default = 20000000) (or a string represented as 200kb or 2MB) indicating the window size for variant labeling. Increase this number for sparser annotation and decrease for denser annotation.
 #' @param nudge_x  A number to vertically adjust the starting position of each gene label (this is a ggrepel parameter)
 #' @param nudge_y  A number to horizontally adjust the starting position of each gene label (this is a ggrepel parameter)
 #' @param angle A number, the angle of the text label
@@ -59,7 +59,7 @@
 #' @param segment.linetype line segment solid, dashed, etc.(ggrepel argument)
 #' @param max.overlaps Exclude text labels that overlap too many things. Defaults to 10 (ggrepel argument)
 #' @param label_fontface  A string or a vector of strings. Label font “plain”, “bold”, “italic”, “bold.italic” (ggrepel argument)
-#' @param label_family A stirng or a vector of strings. Label font name (default ggrepel argument is "")
+#' @param label_family A string or a vector of strings. Label font name (default ggrepel argument is "")
 #' @param gene_label_fontface  Gene label font “plain”, “bold”, “italic”, “bold.italic” (ggrepel argument)
 #' @param gene_label_family Gene label font name (default ggrepel argument is "")
 #' @param build A number representing the genome build. Set to 37 to change to build (GRCh37). The default is build 38 (GRCh38).
@@ -72,6 +72,7 @@
 #' @param vline_alpha The alpha of added vertical line/s (default : 1)  
 #' @param vline_size The size of added vertical line/s (default : 0.5)  
 #' @param region A string representing a genetic region, e.g. chr1:67038906-67359979
+#' @param theme_grey Use gray rectangles (instead of white to distinguish between chromosomes)
 #'
 #' @return ggplot object
 #' @export
@@ -86,16 +87,16 @@
 
 manhattan <- function(df, ntop=4, title="",annotate=NULL, color=get_topr_colors(),
                    sign_thresh=5e-09,sign_thresh_color="red", sign_thresh_label_size=3.5, label_size=3.5, size=0.8,shape=19,alpha=1,
-                   highlight_genes_color="green",highlight_genes_ypos=1,axis_text_size=12,axis_title_size=14, title_text_size=15,
+                   highlight_genes_color="darkred",highlight_genes_ypos=1.5,axis_text_size=12,axis_title_size=14, title_text_size=15,
                    legend_title_size=13,legend_text_size=12, protein_coding_only=TRUE,angle=0,
-                   legend_labels=NULL,chr=NULL, annotate_with="Gene_Symbol",region_size=1000000,legend_name=NULL,
-                   legend_position="bottom", nudge_x=0.1,nudge_y=0.2,xmin=NULL, xmax=NULL,ymin=NULL,ymax=NULL,
+                   legend_labels=NULL,chr=NULL, annotate_with="Gene_Symbol",region_size=20000000,legend_name=NULL,
+                   legend_position="bottom", nudge_x=0.1,nudge_y=0.7,xmin=NULL, xmax=NULL,ymin=NULL,ymax=NULL,
                    highlight_genes=NULL,label_color=NULL,legend_nrow=NULL,gene_label_size=NULL,gene_label_angle=0,
                    scale=1,show_legend=TRUE,sign_thresh_linetype="dashed", sign_thresh_size=0.5,rsids=NULL, rsids_color=NULL,
-                  rsids_with_vline=NULL,annotate_with_vline=NULL,shades_color=NULL,shades_alpha=0.1,segment.size=0.2, 
-                  segment.color="black",segment.linetype="solid",max.overlaps=10,label_fontface="plain",label_family="",
+                  rsids_with_vline=NULL,annotate_with_vline=NULL,shades_color=NULL,shades_alpha=0.5,segment.size=0.2, 
+                  segment.color="black",segment.linetype="dashed",max.overlaps=10,label_fontface="plain",label_family="",
                   gene_label_fontface="plain",gene_label_family="",build=38,verbose=NULL,label_alpha=1,shades_line_alpha=1,vline=NULL,
-                  vline_color="grey",vline_linetype="dashed", vline_alpha=1,vline_size=0.5,region=NULL){
+                  vline_color="grey",vline_linetype="dashed", vline_alpha=1,vline_size=0.5,region=NULL, theme_grey=FALSE){
     top_snps <- NULL
     genes_df <- NULL
     xaxis_label <- "Chromosome"
@@ -103,7 +104,7 @@ manhattan <- function(df, ntop=4, title="",annotate=NULL, color=get_topr_colors(
     dat <- dat %>% set_size_shape_alpha(size, shape, alpha) %>% set_color(color) %>% set_log10p(ntop)
      
     using_ntop <- FALSE
-   
+    
      if(! is.null(region)){
       tmp <- unlist(stringr::str_split(region, ":"))
       chr <- tmp[1]
@@ -129,7 +130,8 @@ manhattan <- function(df, ntop=4, title="",annotate=NULL, color=get_topr_colors(
     }
   if(is.null(ymin)){
     ymin <- get_ymin(dat)
-    if(!is.null(highlight_genes)){ ymin <- ifelse(highlight_genes_ypos < ymin, highlight_genes_ypos, ymin) }
+    if(!is.null(highlight_genes)){ 
+      ymin <- ifelse(highlight_genes_ypos < ymin, highlight_genes_ypos, ymin) }
   }
 
   if(is.null(ymax)){ ymax <- get_ymax(dat) *1.04}
@@ -168,11 +170,10 @@ manhattan <- function(df, ntop=4, title="",annotate=NULL, color=get_topr_colors(
     }
     if(is.null(chr)){
       ticks <- get_ticks(dat)
-         main_plot <- main_plot %>% add_shades_and_ticks(shades,ticks,shades_color=shades_color,shades_alpha=shades_alpha,shades_line_alpha=shades_line_alpha)
+         main_plot <- main_plot %>% add_shades_and_ticks(shades,ticks,shades_color=shades_color,shades_alpha=shades_alpha,shades_line_alpha=shades_line_alpha, theme_grey=theme_grey)
     }else{
       main_plot <- main_plot + scale_y_continuous(expand=c(.02,.02))  + scale_x_continuous(expand=c(.01,.01),labels = scales::comma)
     }
-
   main_plot <- set_axis_labels(main_plot,xaxis_label = xaxis_label)
   main_plot <- main_plot %>% set_plot_text_sizes(axis_text_size=axis_text_size,axis_title_size = axis_title_size, 
                                                  legend_text_size=legend_text_size, legend_title_size=legend_title_size,scale=scale)
@@ -189,7 +190,7 @@ manhattan <- function(df, ntop=4, title="",annotate=NULL, color=get_topr_colors(
 
   if (! is.null(annotate)){
     main_plot <- main_plot %>%  add_annotation(plot_labels = top_snps,annotate_with=annotate_with,angle=angle,label_size = label_size, 
-                                               label_color=label_color, nudge_x=nudge_x, nudge_y=nudge_y, scale=scale, annot_with_vline=annot_with_vline, 
+                                               label_color=label_color, nudge_x=nudge_x, nudge_y=nudge_y, scale=scale, 
                                                segment.size=segment.size,segment.color=segment.color,segment.linetype=segment.linetype, max.overlaps=max.overlaps)
   }
   if (! is.null(highlight_genes)){
@@ -200,7 +201,9 @@ manhattan <- function(df, ntop=4, title="",annotate=NULL, color=get_topr_colors(
                                 label_size=label_size,gene_label_angle = gene_label_angle,scale=scale,gene_label_fontface=gene_label_fontface,gene_label_family=gene_label_family)
     
   }
-
+  if(annot_with_vline){
+    main_plot <- main_plot %>% add_vline(top_snps$POS, vline_color=vline_color, vline_linetype = vline_linetype, vline_alpha=vline_alpha, vline_size=vline_size,scale=scale)
+  }
   with_vline <- FALSE
   if(! is.null(rsids_with_vline)){
     rsids <- rsids_with_vline